Move kubefinder to shared folder
